Modified UI and got some file transferring working
I looked at a previous project I participated in for some UI guidance and restructured the program to achieve a UI I can work with. I had some problems with file transferring and realized it is a linux folder permission issue, I thought I found a solution by transferring the files to friendly directly and using sudo to force it to move but it isn't working for some reason. I'll figure that out later.